Makes GraphExecutor a class instead of DriverV2
Previously we had two drivers, and used a builder pattern to instantiate
them. Now we just have one driver, which enables us to share much more
of the same code. They delegate to the GraphExecutor, which has the
capability to run the graph, and is loosely coupled with components of
the API (E.G. parallelizable[]).