Changes `enable_parallelizable_type` to `enable_dynamic_execution`
After much internal debate we have decided that this is the least
unclear and the least non-descriptive name. Luckily, we can always add
aliases when we come up with a slightly less non-descriptive name.

The truth is that this does a lot, but `enable_v2_execution` was not an
ideal name for us. We will eventually be decoupling the API capabilities
from the executors, which will allow us a better naming scheme
(enable_task_based_execution).