Adding Feast + Hamilton example repository (#222)
This adds an example on how one might use Hamilton together with Feast.

See the README for details, but essentially you'd use Hamilton for defining and running 
feature transformations. You can also use it to help create the Feast Feature object definitions, 
since that can be come a mess to understand. 

All the hard work here was done by @zilto.
Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2.localdomain>