Fixes duplication of node dependencies
This was added when we added graph copying. This adds the ability for
update_dependencies to wipe the dependencies before executing. Note that
it does an in-place modification if this doesn't happen -- that's the
reset_dependencies option. Note this is an internal API so I don't mind
it being geared towards performance.