Adds spark code to run PDF Summarizer
This example shows that you can run the same code in a FastAPI backend,
and then turn around and run it on Spark as well! This is because flows like
PDF summarization are essential “map” transformations, and can then be easily
ported and modeled in spark.

The magic to run on spark happens in the SparkUDFGraphAdapter.

Otherwise, this adds and adjusts code to make the summarization.py code
work on spark (one tweak was needed). We add everything for someone to run the example,  including a script and notebook.

Squashed commit of:

Adds notebook to pdf spark example

So that people can use that to get started too. (+4 squashed commits)
Squashed commits:
[7db102c] Removes unused code.
[ce2abec] Adds missing requirements.txt for spark PDF example

So that people know what to install.
[eae3edd] Adds README to spark PDF post

Renames directory to run_on_spark to make it clear.

Adds pointer to main README to spark one.
[154e3ce] Adds spark code to run PDF Summarizer

We only needed one minor changes to the original code:

* we have to ensure the order of "columns" versus values that can be bound are
in the correct order. So we had to reorder one function's arguments for spark to
work.