Removes strict extension registration
This is for getting it to work with pyspark. While it does have natural
column/dataframe objects, these are not usable in the same way. Thus we
don't want to register the same set of constructs as we would, say, with
polars or pandas.

We allow plugins to opt out by specifying COLUMN_FRIENDLY_DF_TYPE =
False, which defaults to true.