Fixes overrides in dynamic task execution
This is not yet perfect as there are likely a few missing edge cases,
but it fixes the majority of cases that people will encounter. We
wereen't accounting for overrides in task dependencies, so we'd ask for
nodes that did not exist in the DAG. This short-circuits when that
occurs.