Adds defaults to all viz related functions
If you're in a notebook having to pass in None, and empty dicts is cumbersome.
We were also inconsistent across the API. This change makes it so that
output_file_path and render_kwargs have None as defaults. Thereby enabling
someone in a notebook to more ergonomically display an image.