Adds scaffolding for user contributions
We want to enable users to contribute Hamilton dataflows. Data is the oil,
so code isn’t. Therefore we think we can help people get started with Hamilton
by enabling more contributions that enable people to get started quickly with
things off the shelf.

The design premise is to:
1. Have a separate package called `sf-hamilton-contrib` that houses these dataflows.
2. We then autogenerate docusaurus docs based on that module and expose it at hub.dagworks.io.
3. I have created github workflows/and adjusted existing ones to minimize CI usage.
4. This isn’t 100% complete, but I think it’s good enough to be merged into main.

Right now the main thing to note, is the pip install has to be updated with each new sf-hamilton-contrib release until it’s out of RC stage.

Otherwise we're working on a template to ensure things are standardized between dataflows. One major TODO is to figure out the
unit testing story.

Adds to the changelog (+9 squashed commits)
Squashed commits:
[9b5262b] Changes prints to logging statements
[c6e338b] Fixes off by one
[29ef702] Changes commits to be served via static
[0aa1a2d] Fix no official case
[6c84b9b] Adds site map

And upgrades docusaurus.
[1d6bd3e] Adds google analytics to hub

So that we can understand usage.
[6b3a379] Adds telemetry

Fixes a few things too

This commit should be squashed at some point.
[b7e7f7e] Enables gh pages to be run off of main

This is required for things to only run if they are merged to main.
[74a8eee] Adds scaffolding for user contributions

We want to enable users to contribute Hamilton dataflows. Data is the oil,
so code isn’t. Therefore we think we can help people get started with Hamilton
by enabling more contributions that enable people to get started quickly with
things off the shelf.

The design premise is to:
1. Have a separate package called `sf-hamilton-contrib` that houses these dataflows.
2. We then autogenerate docusaurus docs based on that module and expose it at hub.dagworks.io.
3. I have created github workflows/and adjusted existing ones to minimize CI usage.
4. This isn’t 100% complete, but I think it’s good enough to be merged into main.

Cleaning up code a bit (+41 squashed commits)
Squashed commits:
[5aaf10b] changing base to see if it works with custom domain
[0ce0f8d] Fixes build
[a6e828e] Revert "Reverting sidebar"

This reverts commit d21065effda00a809d1ac9bc06bfd3820e3cd404.
[d21065e] Reverting sidebar
[2eff762] FIxing build
[9a5fbaf] Adds more to docs

Adds more structure.
[9867f5e] Adding more docs to contrib module
[fc35509] Adds templates for contrib

You can only have one default. So I decided to link from our default to
the specific one for now. Depending on volume we should switch things around.
[7dc35e2] Constrains build to contrib directory path
[2088037] wip
[a2715f1] Fixes so that pre-commit is run for everything
[58dd21a] Adds more to contrib/readme

should skip tests
[024364d] Fixes script to not error out on non-zero status
[e2c2e78] Adds start of readme

This should stop tests early too.
[2a32754] updates circleci to skip contrib changes
[7fbd056] Test didn't work github manages things
[c7ffdb9] Testing path
[5c780c7] Adds tagging
[0aeb353] reverting path back to what it was
[21f86b1] changes project root
[8ba794c] fixes workflow yaml
[8f2cf89] WIP adjust fetch depth
[953618a] Changes to include all commits
[cc56c44] Fix compile of docs path
[3e954fb] WIP fix python build
[edfe6f6] Adds compile step
[c5dd405] commit to test things
[b215379] Fixes path from moving package
[4826440] wip
[d415a37] WIP packages

Seems to work and gets added to the namespace appropriately without breaking
anything else.
[c1db95f] TODO: fix up package stuff.
[0ffc1b9] WIP docusaurus
[6d08fb7] wip
[1977375] WIP commit
[828060d] WIP fix github action flow
[7e5c3ab] WIP commit
[df5fd77] WIP
[4eb1333] wip
[721df5e] WIP
[9e918b3] wip
[9511774] WIP