Enables the materializers to depend on functions & driver variables
.execute() enables that. The materializers should also be able to handle it.

The complication however is on who knows what and when to do it.
So I made the decision to make the driver sanitize the materializers given
the context that it has. New objects are returned so we don't modify the underlying
materializer object -- that way if someone is reusing a materializer definition they
can do so in other driver runs/etc.