Aggressively prunes Hamilton DAG
This is not perfect, but what we do is after computing and sticking a
node in the memoization dict, we ensure that, for each dependency, if we
know that dependency is not needed by anything else, we delete that from
computed. This way we'll save on memory.

There are likely a few edge cases where this won't work (E.G. we're only
executing a portion of the graph), but in many cases this should help,
and it certainly won't harm.