Splits applies_to for save & loader adapters
Given A --> B

For this directional edge to be valid we want A to be a subclass of what B takes.

For loaders -- we're checking A, the loader,  can be fed into B.  e.g. is loader_A a subclass of input B can take. i.e. the decision function is (loader, function)
For savers -- we're checking A, the function, can be fed into B the saver. e.g.  function_A a subclass of input B the saver can take. i.e. the decision function is (function, loader)

In the code, though, we are treating the two cases the same, as a decision of (loader, function). So this fixes
that by creating two separate functions for a saver and a loader. That way the directional check is correct.

Adds tests for it.