Adds logic to warn when multiple adapters are applicable
Rather than default returning to the last one. We should
check for all, and then return the last one, but warn in the
case of multiple being present. This I think will result in
a better experience. We could make the logging
statement a debug one, but we can change that later.