Refactors converting output values into strings
So both the driver and materializer need to handle converting
function names and variables into strings.

This is a little messy -- but it centralizes logic in common.
I didn't bother with another file because I didn't know what to call it.
So putting under common seems fine.

Otherwise I added tests to ensure that new functionality works,
and left the existing tests to ensure nothing broke.