add sklearn plot saver (#467) implements #456
Adds the ability to save a SKLearn plot as a PNG easily.

Adding plots is a something you may in development and not production,
or vice-versa. So having a this allows one to easily inject this 
at driver time, so you don't have to modify one's logic for concerns
that needn't be hard coded into your dataflow. But there's nothing
wrong with putting it in your dataflow, having it as a materializer
just enables you a little more freedom to decouple concerns.

--- Squashed commits

* add sklear plot saver

* remove unnecessary comment

* Fix review points

* use attrib method

* added tests, fix data saver method

* added test for calibration display

* update req

* handle keyword arguments

* Selective import of display classes

* make compatible with py3.7

* add version checking

* change to 3.8

* add checking to test func

* add additional display classes

* fix failing 3.7 build

* using skipif

* fix signature

* add to plugins_modules list

---------

Co-authored-by: Vivek Praharsha <vpraharsha@outlook.com>
Co-authored-by: vpraharsha03 <vpr03>