Adds validation methods to the driver
This is so you can test it out before you run it. We have
validate_materialization, and validate_execution.

See https://github.com/DAGWorks-Inc/hamilton/issues/484 for more.

There could probably be a slight refactor to drop repeated code, but for
now this is OK.

Included is a validation fix so that it catches the edge-case
effectively of a referenced node not existing in the DAG. This is
grouped together as it is a logically similar change.