Adds path argument to pickle loader
Didn't reallly think about this much, the code is quite ugly when it has
one for loading and one for saving. Also, path is common among every
one, so it should be here. Unfortunately, the easiest way to do this is
to stick it in post_init, which won't be validated till later, but I
think this is an edge case, so I'm not too worried.