Allows for dynamic registration of materializer types
Before we were just registering once. Now we synchrionize all konwn
types if we detect a missing type -- a type that is in the registry but not in the associated class.
Note that we could actually just go to the saver registry
directly, but we like to add to it dynamically so we can refer to it,
say, in a jupyter notebook. If we don't add it dynamically, we'll never
know the available set.

See #521