Adds duplicate contrib/__init__ for telemetry capture
So if someone does not have sf-hamilton-contrib installed, you could
download a dataflow but not be able to import it, because they
have a dependency on having a `hamilton.contrib` submodule. This sticks
a placeholder in -- to enable things to work without installing sf-hamilton-contrib.
The assumption is that it will get clobbered by installed `sf-hamilton-contrib`,
and everything should continue to work.

This also updates the telemetry code to handle this case, and to parse
the relevant details from the file path. I have tested that things work
for old and new paths.

What I haven't extensively tested is the installation order.