Various hub/contrib improvements (#528)
At a high level this commit does:
 - moves things around, e.g. official -> DAGWorks.
 - fixes copy and READMEs
 - adds stub for leaderboard
 - fixes telemetry/contrib module usage when sf-hamilton-contrib is not installed



See squashed commits for details:

* Moves Official to DAGWorks for contrib/hub

Because that's who is maintaining it.

I also tacked on refactoring the python hub docs
code into jinja2 to enable more easily having
conditional logic that is easier to maintain.

* Adds "view by tags" link in hub

So that people can see an index.

This is a quick way to get to this screen
without really trying to customize docusaurus.

* Fixes some README wording to drop `==` install

For the hub we don't want to have to bump the version
in the docs. So removing specifying the version for now.
Though we could build something to know what the version
is that's more effort than I want to do at the moment.

* Adds link to copy docs for hub

* Adds duplicate contrib/__init__ for telemetry capture

So if someone does not have sf-hamilton-contrib installed, you could
download a dataflow but not be able to import it, because they
have a dependency on having a `hamilton.contrib` submodule. This sticks
a placeholder in -- to enable things to work without installing sf-hamilton-contrib.
The assumption is that it will get clobbered by installed `sf-hamilton-contrib`,
and everything should continue to work.

This also updates the telemetry code to handle this case, and to parse
the relevant details from the file path. I have tested that things work
for old and new paths.

What I haven't extensively tested is the installation order.

* Moves polars import to try/except

For some reason this was breaking for me,
and this fixed it.

* Fixing isort issues

* Cleaning up some typos

* Revert "Fixing isort issues"

This reverts commit 8288b7ffadfd705a50fa7ec412828ca838946466.

* Adds stub for leaderboard page for hub

This is pretty basic. But putting something up here.

* Fixing isort