Supports "Any" in loader types
There are certain cases where the loader can load any datatype , E.G.
pickle-type loaders. Before this didn't work. The "proper" solution
would be to change custom_subclass_check, but that has the possibility
of breaking some cases of materialization as we have not always been
consistent here. In this case, it *just* uses a loader with `Any` if
another has not been specified. While this might casue the order to be
slightly counter-intuitive, it will ensure full backwrads compatibility.