Fixes parallelize when parameterization is used
We weren't copying over the node_role. This does so + adds tests.
Note there's something strange happening with copy_with -- it breaks
when we have dependencies copied earlier. I think this is because we're
reusing nodes, but I"m not sure. At some point we'll want to dig in, but
for now this is a good enough fix.