Fixes version check
There is something funky going on. This only just suddenly started to break.
So it could be we were just lucky and now some installation non-determinism
is throwing things off, or something else changed...

Otherwise this hacks around the case of `pip install -e .` not doing the
correct thing due the package structure we've set up.