Fixes parallel case with empty list
This was broken as we assumed the list was not empty. This special-cases
it. We should probably be breaking this into functions so we can add
more unit tests, but given the experimental nature of this I'm not
worried. Don't want to over-engineer. The fix will work nicely for now.