Attaches extra data with @ config resolution
So we lost the original function name when resolving. This
attaches it as an attribute along with the configuration keys.

We then attach the keys used as tags on the node, and
otherwise changes the error message to pull out the original
function name.

The error message code is coupled to how @config resolution works. Ideally it shouldn't be.
So when @config resolvers are changed to return Nodes, then fn.__name__ should
just work in the error part.