Schema management for dataset-types
As discussed, we're going with a list of tuples, and not validating the types.
That said, we have a set of types we allow and we can assume that any of those are supported,
whereas the others are allowed but will not be fully supported.

Note we also made the following decisions:

1. We implement the types as a single tag internally. This will change,
   so its not exposed to the user
2. We only allow this to decorate registered dataframe types