Makes publicly available HamiltonNode/HamiltonGraph classes
We plan to use these in multiple place for user-facing APIs.
HamiltonNode replaces the old Variable, which it leaves behind in Driver
for backwards compatibility. TBD whether we want to have the two
concepst that mean the same thing or unit them fully, but we can decide
that later, as they point to the same thing.