Fixes circular imports
There's some oddities here with graph/dependencies. The problem is that
we mixed together a bunch of abstractions for Graph adaptesr, which when
unbundling we hit a few circular imports. Luckily these were purely in
implied types (although its a bit too close for comfort to actual
logic), so we can add an if TYPE_CHECKING statement and we should be
good to go. THat said, we'll want to revisit -- see notes in code as to
better approaches.