Moves post-task-execute into finally block
This is functionally the same as before, it just hits a weird edge case
in which a failure to deliver the hook would be called twice, as it
would fail when running it and fail when reacting to the failure (in the
catch). We need to do this for graph/node.