Adds datadog plugin with tracing support
This adds basic tracing support for the datadog plugin. We elect to
manage the state of spans/traces ourselves, and there is a decent amount
of complexity around serialization. It works on Ray, Dask,
Multithreading, and Synchronous mode, as well as vanilla Hamilton
(non-task-based). It can optionally store links, although datadogs
functionality is pretty suboptimal and has large UI query times,
so that is turned off by default.