Fixes all local file loader metadata to have a uniform shape
Beforehand we had different approaches depending on whether it was a
dataframe. We'll need this for downstream parsing. The new format is one
to many of:
1. sql_metadata
2. file_metadata
3. dataframe_metadata

The utils functions will return prefixed dicts, and consumers just have
to add them. This is all internal-facing (just to be used for
diagnostics), so its OK to change it.