Adds fine tuning contrib example (#667)
* Adds fine tuning contrib example

This module shows how one can perform fine tuning using
Hamilton. It is a basic example using the transformers
library that connects to huggingface to pull and fine tune
a FLAN model.

One should be able to adapt this code to their needs.

* Updates to PR from feedback.

* Adds extra comments for first tokenization functions

* Updates docker file comment

So that people know to change it to match
their dataset.

* Bumps version to get 0.0.7 out

With latest contrib additions.