Fixes tag_outputs to tag everything
This was overly restrictive at first. tag_outputs should be able to tag
any intermediate nodes as well, given that its specified by name. We use
the ... annotation to say that we want to tag everything, but that's
hidden internally. While this could be construed as backwards
incompatible, it would be a very strange case, and its more reasonable
to argue that the documentation suggests that this is the intendecd
behavior and it was always wrong.