Docs: large update to the `concepts/` section (#702)
`docs/concepts` pages updated

pages edited / added
- builder
- driver
- function modifiers
- materialization
- visualization
- lexicon / glossary

Concepts should be introduced with bite-size snippets that highlight the core notion without trying to be exhaustive. For example, the introductory Driver page simply mentions loading a module and executing/viewing the dataflow. 

The learning experience can be improved by providing brief use case examples, such as using `@config.when` to swap model type. Side-by-side display of code or pictures helps build a mental model of the concept and highlights the value of a given feature.

This is part of a greater documentation overhaul to improve the user journey / onboarding. 

Some deprecated examples and code snippets were updated.

---------

Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2>
Co-authored-by: Stefan Krawczyk <stefan@dagworks.io>