Adds example on how to chunk Hamilton documentation
This uses langchain's text splitting functionality, you
could easily replace this with your own logic. The
DAG though is run and defined in Hamilton -- and the processing
of each URL is done in parallel. You could easily farm
this out to Ray or Dask as well... See the documentation
for more info.