Fixes bugs in pyspark @with_columns
1. Configuration linkage was not exposed. So adding a parameter
for people to specify what configuration is required.
2. There was a bug in the UDF logic, because `all(empty)` evaluates
to True. So fixed that.
3. The `select` in append mode appended all possible columns. It
didn't only append what was in the `select`. This changes that behavior
by dropping "outputs" from the subdag that aren't in the select. We
do it this way because we don't know what's in the original dataframe.
4. Adds test to check for the new logic