Typer CLI refactor and `hamilton validate` addition
Bit of scope creep.

Changes:
- drop support .ini config because they don't have
a basic construct for lists, which is required for
final_vars in driver.validate_execution()
- rename config methods to context. It allows to
reduce the number of arguments to pass to the CLI
and containts Driver config + execution final_vars,
inputs, and overrides
- refactored the CLI to extract methods to try a
command and to display a result in terminal.
- added a validate command. While build checks for
a valid definition, validate checks that certain
dataflow paths aren't broken for a given context.