Better pipeline layout to work with with_columns
This is close to the original, but we need to specify
the set of spark safe functions. This is okay-ish I think.
It just also requires that we include the pipeline in the
larger DAG too... (+1 squashed commit)
Squashed commits:
[fe086461] Adds code to do document chunking with pyspark

with_columns makes some assumptions:
(a) that all functions in the with_columns subdag take in the dataframe
(b) that all intermediate functions in the subdag need to become columns in the dataframe
(c) it doesn't allow you to wire through inputs to the subdag

We'll need to improve upon some of these to make using
 spark a real drop in replacement.

The most annoying part right now is not being able to wire through
inputs to the subdag from other functions in the DAG. That's why
the html_chunker and text_chunker are outside Hamilton right now.