Adds RAG example on how to chunk Hamilton documentation (#721)
This is the first part of a RAG pipeline - document ingestion.
This uses langchain's text splitting functionality, you
could easily replace this with your own logic. The
DAG though is run and defined in Hamilton -- and the processing
of each URL is done in parallel. You could easily farm
this out to Ray or Dask, and even pyspark as well... See the README
for more info.