Adds FunctionInputOutputTypeChecker
This is a lifecycle adapter that one can add to strictly enforce the types
of things at runtime. It will check the input to and the output from a function
and check that it matches the expected type annotation.

It modifies/fixes the pandas split-apply-combine example to conform
to the expected types; it actually found mismatches between what was
typed and what was passed in.

I had to modify graph_functions.py so that the error message from the
pre-node-execute hook would show up in the right place and the
post-node-execute hook would not run.

Note: this does not do anything smart with Pandera or Annotated types
just yet, but it could in future.

Test generated with the help of co-pilot. A little ugly, but they do the
job.