Adds originating functions for materializers and changes version logic
HamiltonNode.version was returning None in certain cases. This
broke graph.version. Culprits: config, and materializers.

This change:

1. makes the node version for config == name of the node.
2. makes materializers all have originating function values.
3. guards graph.version against computing functions with None
as the version value.

So node.version should now hardly ever be None. But if it is
for some strange reason, we should now handle it better.

Adds tests for it.
Also fixed a fixture that was failing for me for the shelve caching
adapter.