Fixing hashing for 3.8
So different python versions hash differently.
So rather than doing a lot of work to figure out the code
and hash of each node... I am hardcoding the hash for
versions that are different. It appears that 3.8 is the odd
one out -- all the other python versions seem to hash
to the same thing.