Adds guards around version logic (#789)
* Adds guards around version logic

HamiltonNode.version was returning None in certain cases. This
broke graph.version. Culprits: config, and sometimes materializers.

This change:

1. makes the node version for config == name of the node.
2. guards graph.version against computing functions with None
as the version value.
3. after investigating more, materializers aren't the issue. We could
include the materializer code in the originating function if needed though.

So node.version should now hardly ever be None. But if it is
for some strange reason, we should now handle it better.

Adds tests for it.
Also fixed a fixture that was failing for me for the shelve caching
adapter.

* Fixing hashing for 3.8

So different python versions hash differently.
So rather than doing a lot of work to figure out the code
and hash of each node... I am hardcoding the hash for
versions that are different. It appears that 3.8 is the odd
one out -- all the other python versions seem to hash
to the same thing.

* Reverting adding originating function to materializers

We can figure something better out later. Right now what
I had worked, but wasn't useful.