Adds Ray Remote Options Support (#800)
Adds Ray remote options support.

Thanks to Fran Boon for work!

Initializes a `@ray_remote_options` decorator. This takes in a list of options to pass to ray.remote().
Supported options include resources, as well as other options:
 - https://docs.ray.io/en/latest/ray-core/scheduling/resources.html

This is implemented using tags, but that might change. Thus you should not directly
rely on the tags created by this decorator (which is why they are on a reserved namespace).

To use this, you should decorate a node with `@ray_remote_options`
Example usage:

```python

@ray_remote_options(
      num_gpus=1,
      resources={"my_custom_resource": 1},
)
def example() -> pd.DataFrame:
      ...
```
---------
Co-authored-by: Fran Boon <francis.boon@oxehealth.com>