Implements Hamilton UI -- OS from previously closed source (#840)
* Implements Hamilton UI -- OS from previously closed source

This has the following components:
1. Frontend -- React + typescript app that works together with the
   server
2. Backend -- Django app connected to a postgres DB + blob store.
   Configurable. Defaults to no authentication in local mode, but can be
   connected with PropelAuth.
3. Deployment -- tools for docker-compose deployment
4. Documentation -- minimally documented for now but has enough to get
   started

Note that this does not yet enable tests/github actions, that will be
done in the next few steps. That said, they exist for now.

* Runs/fixes pre-commit

* Adds backend tests to github actions

We're using github actions so we can conditionally skip it.

* Adds frontend github actions

This just tests a build/lint process. We're using github actions for all
new ones so we can do conditional/iterate faster.

* Simplifies github actions for UI

Github actions can filter on changes based on directory path.

---------

Co-authored-by: Stefan Krawczyk <stefan@dagworks.io>
Co-authored-by: Elijah ben Izzy <elijah@dagworks.io>