Adds basic langchain support for SDK
This covers documents and messages. It is basic, and not
mean to really enable deserialization, but we do have the
components here to know which is which.

Otherwise rather than a special type, using dict for now. Can
always change it and coordinate with the UI.

This also adds missing tests for the other new SDK additions.

Note: we use `_stats` as the prefix for these files to indicate that there could
be missing information. If we wanted to properly cache etc,
we could do so, but we'd want to do it via another means I think.