Adds Union type check for saver if it supports multiple types
Say we have this - and want to save it with a saver:

```python
def foo() -> Union[int, float]:
   return ...
```

If the saver's applicable_types is [int, float], this would previously fail, now
it does not. Added test for this.

If the saver's applicable_type was just `float` or `int`, then rightly this fails -- added
test for that explicitly.