added `__getitem__()` and `filter_nodes()`to `HamiltonGraph` (#919)
* added `__getitem__()` and `filter_nodes()` to `HamiltonGraph`

The most common operations on the HamiltonGraph is selecting a single node or filtering them by a criteria. However, they're are no API for that, meaning some list comprehension or filter() are often implemented in the library code but also in the user's code. This motivated the implementation __getitem()__ and filter_nodes()

---------

Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2>