Enables inputs to be also outputs; Fixes #936
This enables one to pass in inputs and request them as outputs
independent of the graph.

The use case here is that you want to join some data at the end
that is extra and not in the DAG. E.g. extra pandas data.

Adds test cases to check for different angles.